Interchanged tabs between left and right panel. Will continue to monitor effects of this change and test. For now, it is functional
